// src/generated/proto/products.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.12.4
//   protoc               v3.19.1
// source: products.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "products";

export interface ProductById {
  id: number;
}

export interface CreateProductRequest {
  name: string;
  price: number;
}

export interface UpdateProductRequest {
  id: number;
  name?: string | undefined;
  price?: number | undefined;
}

export interface PaginationRequest {
  page?: number | undefined;
  limit?: number | undefined;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  available: boolean;
}

export interface PaginationMeta {
  total: number;
  page: number;
  lastPage: number;
}

export interface ProductList {
  data: Product[];
  meta: PaginationMeta | undefined;
}

export interface ValidateProductsRequest {
  ids: number[];
}

export interface ValidateProductsResponse {
  data: Product[];
}

export const PRODUCTS_PACKAGE_NAME = "products";

export interface ProductsServiceClient {
  create(request: CreateProductRequest): Observable<Product>;

  findAll(request: PaginationRequest): Observable<ProductList>;

  findOne(request: ProductById): Observable<Product>;

  update(request: UpdateProductRequest): Observable<Product>;

  remove(request: ProductById): Observable<Product>;

  validateProducts(request: ValidateProductsRequest): Observable<ValidateProductsResponse>;
}

export interface ProductsServiceController {
  create(request: CreateProductRequest): Promise<Product> | Observable<Product> | Product;

  findAll(request: PaginationRequest): Promise<ProductList> | Observable<ProductList> | ProductList;

  findOne(request: ProductById): Promise<Product> | Observable<Product> | Product;

  update(request: UpdateProductRequest): Promise<Product> | Observable<Product> | Product;

  remove(request: ProductById): Promise<Product> | Observable<Product> | Product;

  validateProducts(
    request: ValidateProductsRequest,
  ): Promise<ValidateProductsResponse> | Observable<ValidateProductsResponse> | ValidateProductsResponse;
}

export function ProductsServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["create", "findAll", "findOne", "update", "remove", "validateProducts"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("ProductsService", method)(constructor.prototype[method], method, descriptor);
      Object.defineProperty(constructor.prototype, method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("ProductsService", method)(constructor.prototype[method], method, descriptor);
      Object.defineProperty(constructor.prototype, method, descriptor);
    }
  };
}

export const PRODUCTS_SERVICE_NAME = "ProductsService";